resolution of celluloid/culture#31 for Celluloid::IO
