+++ spec/celluloid/io/actor_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
 
-describe Celluloid::IO do
+RSpec.describe Celluloid::IO, library: :IO do
   it_behaves_like "a Celluloid Actor", Celluloid::IO
 end
