+++ spec/celluloid/io/dns_resolver_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-describe Celluloid::IO::DNSResolver do
-  describe '#resolve' do
+RSpec.describe Celluloid::IO::DNSResolver, library: :IO do
+  context '#resolve' do
     it 'resolves hostnames statically from hosts file without nameservers' do
       # /etc/resolv.conf doesn't exist on Mac OSX when no networking is
       # disabled, thus .nameservers would return nil
