+++ spec/celluloid/io/mailbox_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
 
-describe Celluloid::IO::Mailbox do
+RSpec.describe Celluloid::IO::Mailbox, library: :IO do
   it_behaves_like "a Celluloid Mailbox"
 end
