+++ spec/celluloid/io/reactor_spec.rb
@@ -1,19 +1,19 @@
 require 'spec_helper'
 
-describe Celluloid::IO::Reactor do
+RSpec.describe Celluloid::IO::Reactor, library: :IO do
   let(:payload) { "balls" }
   let(:example_port) { assign_port }
-  
+
   it "shouldn't crash" do
     server = ::TCPServer.new example_addr, example_port
-    
+
     thread = Thread.new { server.accept }
-    
+
     socket = within_io_actor { Celluloid::IO::TCPSocket.new example_addr, example_port }
     peer = thread.value
     peer_thread = Thread.new { loop { peer << payload } }
     handle = false
-    
+
     # Main server body:
     within_io_actor do
       begin
@@ -29,9 +29,9 @@ describe Celluloid::IO::Reactor do
         handle = true
       end
     end
-    
+
     expect(handle).to be_truthy
-    
+
     server.close
     peer.close
     socket.close
