+++ spec/celluloid/io_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Celluloid::IO do
+RSpec.describe Celluloid::IO, library: :IO do
   context "copy_stream" do
     let(:host) { "127.0.0.1" }
     let(:port) { 23456 }
